backend/requirements.txt
Flask
Flask-Cors
tensorflow-cpu
numpy
opencv-python
gunicorn
